requirements.txt
docker==6.1.3
feature-engine==1.6.1
hyperopt==0.2.7
joblib==1.3.2
matplotlib==3.7.2
mlflow==2.6.0
numpy==1.24.4
pandas==2.0.0
pydantic==1.10.9
pytest==7.4.2
PyYAML==6.0.1
scikit-learn==1.3.0
seaborn==0.12.2
shap==0.42.1
SQLAlchemy==2.0.20
tqdm==4.66.1
xgboost==1.7.6
fastapi==0.103.1
loguru==0.7.2
uvicorn==0.20.0
